+++ Sources/Turf/Geometry.swift
@@ -1,0 +1,24 @@
+import Foundation
+#if !os(Linux)
+import CoreLocation
+#endif
+
+
+public enum GeometryType: String {
+    case Point
+    case LineString
+    case Polygon
+    case MultiPoint
+    case MultiLineString
+    case MultiPolygon
+    
+    static let allValues: [GeometryType] = [.Point, .LineString, .Polygon, .MultiPoint, .MultiLineString, .MultiPolygon]
+}
+
+public struct Geometry: Codable {
+    public var type: String
+    
+    public var geometryType: GeometryType? {
+        return GeometryType(rawValue: type)
+    }
+}
